pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "seam-lint"
version = "0.2.0"
description = "Coherence fee diagnostic for agent tool compositions"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Res Agentica" }]
keywords = ["mcp", "agents", "linter", "composition", "coherence"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = ["pyyaml>=6.0"]

[project.scripts]
seam-lint = "seam_lint.cli:main"

[project.urls]
Homepage = "https://github.com/jkomkov/seam-lint"
Issues = "https://github.com/jkomkov/seam-lint/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/seam_lint"]

[tool.pytest.ini_options]
testpaths = ["tests"]
